- Move Security to http folder where it belongs - Make sure all models extend EasyPostObject - Fix HTTP method for shipment refund function
